Cleanup custom acquisition adapters
